Added documentation Re #24748
